requirements.txt
asgiref @ file:///home/conda/feedstock_root/build_artifacts/asgiref_1642912811357/work
autopep8 @ file:///home/conda/feedstock_root/build_artifacts/autopep8_1635267974115/work
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django @ file:///home/conda/feedstock_root/build_artifacts/django_1617829946528/work
django-allauth==0.50.0
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
idna==3.3
oauthlib==3.2.0
pycodestyle @ file:///home/conda/feedstock_root/build_artifacts/pycodestyle_1633982426610/work
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1647961439546/work
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse @ file:///home/conda/feedstock_root/build_artifacts/sqlparse_1631317292236/work
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1604308577558/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1644850595256/work
urllib3==1.26.9
